requirements.txt
fastapi
uvicorn[standard]
google-genai
pydantic
load_dotenv